item

+++ app/src/main/java/Host/SimpleDamageCalculator.java
@@ -5,6 +5,6 @@ import Common.Player;
 public class SimpleDamageCalculator implements DamageCalculator {
     @Override
     public int calculateDamage(Player attacker, long ms) {
-        return attacker.getProperty().getDPS() * ms / 1000;
+        return (int) (attacker.getProperty().getDPS() * ms / 1000);
     }
 }
